dfxvideo: Readded windows support.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@56047 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfxvideo/prim.c
@@ -387,9 +387,9 @@ void cmdTexturePage(unsigned char * baseAddr)
 
  lGPUstatusRet&=~0x000007ff;
  lGPUstatusRet|=(gdata & 0x07ff);
- 
- usMirror=gdata&0x3000;
- 
+
+ usMirror=(unsigned short)(gdata&0x3000);
+
  UpdateGlobalTP((unsigned short)gdata);
  GlobalTextREST = (gdata&0x00ffffff)>>9;
 }
